requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2
django-cors-headers==3.10.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
gitdb==4.0.9
GitPython==3.1.27
kombu==5.2.3
openpyxl==3.0.9
packaging==21.3
prompt-toolkit==3.0.28
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2021.3
redis==4.1.4
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
vine==5.0.0
watchdog==2.1.6
wcwidth==0.2.5
wrapt==1.13.3
